requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-pubsub==2.18.4
ollama==0.1.7

